Fix static build with MinGW on Linux
The static MinGW Qt 5 available in Fedora's repositories doesn't import the QWindows plugin automatically. This results in Windows being unable to start the program, since the plugin is missing.
The solution is to add the required libraries in the project file and manually import the QWindows plugin in main.cpp.