Plugins.cpp: use QStandardPaths::HomeLocation (Qt 5) and QDesktopServices::HomeLocation (Qt 4)
QStandardPaths:AppDataLocation has been added in Qt 5.4, but Ubuntu 14.04 (the latest available on Travis CI at the moment) has Qt 5.2.1.

To overcome this problem, we can use QStandardPaths::HomeLocation to retrieve the user's home folder and then manually append the rest of the path.

This commit also adds support for Qt 4, by checking Qt's version and then including and using QStandardPaths or QDesktopServices according to it.